+++ dotfiles.powershell/Public/aliases.ps1
@@ -1,0 +1,3 @@
+# my aliases
+
+New-Alias -Name gmnav -Value Get-ModuleNounsAndVerbs